revamped convertors, added lazy class initialisations to Module, fixes pickling issues
