+++ ivy/stateful/converters.py
@@ -3,11 +3,10 @@
 from typing import Optional, Dict, List
 import re  # noqa
 import inspect
-from collections import OrderedDict
 
 # local
 import ivy
-from ivy.functional.ivy.gradients import _is_variable
+
 from ivy.utils.backend import current_backend
 
 
@@ -119,72 +118,12 @@ class ModuleConverters:
             )
 
         try:
-            from haiku._src.data_structures import FlatMapping
+            from haiku._src.data_structures import FlatMapping  # noqa
         except (ImportError, AttributeError):
             raise ImportError(
                 "Unable to import `FlatMapping` from `haiku`. Please check if the "
                 "requested attribute exists."
             )
-
-        def _hk_flat_map_to_dict(hk_flat_map):
-            ret_dict = dict()
-            for k, v in hk_flat_map.items():
-                new_k = k.replace("/", "|")
-                if isinstance(v, FlatMapping):
-                    ret_dict[new_k] = _hk_flat_map_to_dict(v)
-                else:
-                    ret_dict[new_k] = v
-            return ret_dict
-
-        def _dict_to_hk_flat_map(dict_in):
-            ret_flat_map = dict()
-            for k, v in dict_in.items():
-                new_k = k.replace("|", "/")
-                if isinstance(v, dict):
-                    ret_flat_map[new_k] = _dict_to_hk_flat_map(v)
-                else:
-                    ret_flat_map[new_k] = v
-            return FlatMapping(ret_flat_map)
-
-        class HaikuIvyModule(ivy.Module):
-            def __init__(
-                self, *args, params_hk, native_module, device, devices, **kwargs
-            ):
-                self._native_module = native_module
-                self._args = args
-                self._kwargs = kwargs
-                ivy.Module.__init__(
-                    self,
-                    params_hk,
-                    *args,
-                    build_mode="on_init",
-                    device=device,
-                    devices=devices,
-                    **kwargs,
-                )
-
-            def _create_variables(self, device, dtype):
-                return self._hk_params
-
-            def _build(self, params_hk, *args, **kwargs):
-                args, kwargs = ivy.args_to_native(*args, **kwargs)
-                # noinspection PyUnresolvedReferences
-                params_dict = _hk_flat_map_to_dict(params_hk)
-                self._hk_params = ivy.Container(params_dict, dynamic_backend=False)
-                param_iterator = self._hk_params.cont_to_iterator()
-                _, param0 = next(param_iterator, ["_", 0])
-                if hasattr(param0, "device"):
-                    self._dev = ivy.as_ivy_dev(param0.device())
-                else:
-                    self._dev = ivy.as_ivy_dev("cpu")
-
-            def _forward(self, *a, **kw):
-                a, kw = ivy.args_to_native(*a, **kw)
-                params_hk = _dict_to_hk_flat_map(self.v.cont_to_dict())
-                ret = self._native_module.apply(params_hk, 0, *a, **kw)
-                if isinstance(ret, tuple):
-                    return ivy.args_to_native(*ret)
-                return ivy.to_native(ret)
 
         c_args = ivy.default(constructor_args, [])
         c_kwargs = ivy.default(constructor_kwargs, {})
@@ -206,8 +145,9 @@ class ModuleConverters:
 
             transformed_module = hk.transform(forward_fn)
             params_hk = transformed_module.init(rng_seed, *i_args, **i_kwargs)
+        from ivy.stateful.module import _HaikuIvyModule
 
-        return HaikuIvyModule(
+        return _HaikuIvyModule(
             *i_args,
             params_hk=params_hk,
             native_module=transformed_module,
@@ -264,7 +204,7 @@ class ModuleConverters:
             The new trainable ivy.Module instance.
         """
         try:
-            import flax
+            import flax  # noqa
         except ModuleNotFoundError:
             raise ModuleNotFoundError(
                 "`flax` was not found installed on your system. Please proceed "
@@ -278,43 +218,6 @@ class ModuleConverters:
                 "`jax` was not found installed on your system. Please proceed "
                 "to install it and restart your interpreter to see the changes."
             )
-
-        class FlaxIvyModule(ivy.Module):
-            def __init__(
-                self, *args, params_fx, native_module, device, devices, **kwargs
-            ):
-                self._native_module = native_module
-                self._args = args
-                self._kwargs = kwargs
-                ivy.Module.__init__(
-                    self,
-                    params_fx,
-                    *args,
-                    build_mode="on_init",
-                    device=device,
-                    devices=devices,
-                    **kwargs,
-                )
-
-            def _create_variables(self, device, dtype):
-                return self._fx_params
-
-            def _build(self, params_fx, *args, **kwargs):
-                args, kwargs = ivy.args_to_native(*args, **kwargs)
-                # noinspection PyUnresolvedReferences
-                params_dict = flax.core.unfreeze(params_fx)
-                self._fx_params = ivy.Container(params_dict, dynamic_backend=False)
-                param_iterator = self._fx_params.cont_to_iterator()
-                _, param0 = next(param_iterator, ["_", 0])
-                self._dev = ivy.as_ivy_dev(ivy.dev(param0))
-
-            def _forward(self, *a, **kw):
-                a, kw = ivy.args_to_native(*a, **kw)
-                params_fx = flax.core.freeze(self.v.cont_to_dict())
-                ret = self._native_module.apply(params_fx, *a, **kw)
-                if isinstance(ret, tuple):
-                    return ivy.args_to_native(*ret)
-                return ivy.to_native(ret)
 
         c_args = ivy.default(constructor_args, [])
         c_kwargs = ivy.default(constructor_kwargs, {})
@@ -334,8 +237,9 @@ class ModuleConverters:
 
             native_module = native_module(*c_args, **c_kwargs)
             params_fx = native_module.init(rng_seed, *i_args, **i_kwargs)
+        from ivy.stateful.module import _FlaxIvyModule
 
-        return FlaxIvyModule(
+        return _FlaxIvyModule(
             *i_args,
             params_fx=params_fx,
             native_module=native_module,
@@ -383,40 +287,6 @@ class ModuleConverters:
         ret
             The new trainable ivy.Module instance.
         """
-
-        class KerasIvyModule(ivy.Module):
-            def __init__(self, *args, native_module, device, devices, **kwargs):
-                self._native_module = native_module
-                self._args = args
-                self._kwargs = kwargs
-
-                ivy.Module.__init__(
-                    self, *args, device=device, devices=devices, **kwargs
-                )
-
-            def _create_variables(self, device=None, dtype=None):
-                return self._native_params
-
-            def _build(self, *args, **kwargs):
-                self._native_params = ivy.Container(
-                    OrderedDict(
-                        sorted(
-                            [
-                                (param.name, param)
-                                for param in self._native_module.variables
-                            ]
-                        )
-                    ),
-                    dynamic_backend=False,
-                )
-
-            def _forward(self, *a, **kw):
-                a, kw = ivy.args_to_native(*a, **kw)
-                ret = self._native_module(*a, **kw)
-                if isinstance(ret, tuple):
-                    return ivy.args_to_native(*ret)
-                return ivy.to_native(ret)
-
         c_args = ivy.default(constructor_args, [])
         c_kwargs = ivy.default(constructor_kwargs, {})
         i_args = ivy.default(instance_args, [])
@@ -431,8 +301,9 @@ class ModuleConverters:
             native_module = native_module(*c_args, **c_kwargs)
             input_shape = i_args[0].shape
             native_module.build((input_shape[-1],))
+        from ivy.stateful.module import _KerasIvyModule
 
-        return KerasIvyModule(
+        return _KerasIvyModule(
             *i_args,
             native_module=native_module,
             device=device,
@@ -479,42 +350,6 @@ class ModuleConverters:
         ret
             The new trainable ivy.Module instance.
         """
-
-        class PaddleIvyModule(ivy.Module):
-            def __init__(self, *args, native_module, device, devices, **kwargs):
-                self._native_module = native_module
-                self._args = args
-                self._kwargs = kwargs
-
-                ivy.Module.__init__(
-                    self, *args, device=device, devices=devices, **kwargs
-                )
-
-            def _create_variables(self, device=None, dtype=None):
-                return self._native_params
-
-            def _build(self, *args, **kwargs):
-                self._native_params = ivy.Container(
-                    OrderedDict(
-                        sorted(
-                            [
-                                (k.replace(".", "/"), v)
-                                for k, v in dict(
-                                    self._native_module.named_parameters()
-                                ).items()
-                            ]
-                        )
-                    ),
-                    dynamic_backend=False,
-                )
-
-            def _forward(self, *a, **kw):
-                a, kw = ivy.args_to_native(*a, **kw)
-                ret = self._native_module(*a, **kw)
-                if isinstance(ret, tuple):
-                    return ivy.args_to_native(*ret)
-                return ivy.to_native(ret)
-
         c_args = ivy.default(constructor_args, [])
         c_kwargs = ivy.default(constructor_kwargs, {})
         i_args = ivy.default(instance_args, [])
@@ -522,8 +357,9 @@ class ModuleConverters:
 
         if inspect.isclass(native_module):
             native_module = native_module(*c_args, **c_kwargs)
+        from ivy.stateful.module import _PaddleIvyModule
 
-        return PaddleIvyModule(
+        return _PaddleIvyModule(
             *i_args,
             native_module=native_module,
             device=device,
@@ -575,83 +411,12 @@ class ModuleConverters:
             The new trainable ivy.Module instance.
         """
         try:
-            import torch
+            import torch  # noqa
         except ModuleNotFoundError:
             raise ModuleNotFoundError(
                 "`torch` was not found installed on your system. Please proceed "
                 "to install it and restart your interpreter to see the changes."
             )
-
-        class TorchIvyModule(ivy.Module):
-            def __init__(
-                self, *args, native_module, device, devices, inplace_update, **kwargs
-            ):
-                self._native_module = native_module
-                self._args = args
-                self._kwargs = kwargs
-                self._update_v = (
-                    self._inplace_update_v if inplace_update else self._replace_update_v
-                )
-                ivy.Module.__init__(
-                    self, *args, device=device, devices=devices, **kwargs
-                )
-
-            def _create_variables(self, device=None, dtype=None):
-                return self._native_params
-
-            def _build(self, *args, **kwargs):
-                self._native_params = ivy.Container(
-                    OrderedDict(
-                        sorted(
-                            [
-                                (k.replace(".", "/"), v)
-                                for k, v in dict(
-                                    self._native_module.named_parameters()
-                                ).items()
-                            ]
-                        )
-                    ),
-                    dynamic_backend=False,
-                )
-
-            @staticmethod
-            def _inplace_update(p, v):
-                p.data = v.data
-
-            def _inplace_update_v(self, new_v):
-                ivy.Container.cont_multi_map(
-                    lambda xs, kc: self._inplace_update(xs[0], xs[1]),
-                    [self._native_params, new_v],
-                )
-
-            def _replace_update_v(self, new_v, native=None):
-                native = ivy.default(native, self._native_module)
-                for k, v in new_v.items():
-                    if isinstance(v, ivy.Container):
-                        # noinspection PyProtectedMember
-                        native._modules[k] = self._replace_update_v(
-                            v, native._modules[k]
-                        )
-                    elif _is_variable(v):
-                        # noinspection PyProtectedMember
-                        native.__setattr__(k, v)
-                    elif isinstance(v, torch.Tensor):
-                        # noinspection PyProtectedMember
-                        native.__setattr__(k, torch.nn.Parameter(v))
-                    else:
-                        raise ivy.utils.exceptions.IvyException(
-                            "found item in variable container {} which was neither a "
-                            "sub ivy.Container nor a variable.".format(v)
-                        )
-                return native
-
-            def _forward(self, *a, **kw):
-                a, kw = ivy.args_to_native(*a, **kw)
-                self._update_v(self.v)
-                ret = self._native_module(*a, **kw)
-                if isinstance(ret, tuple):
-                    return ivy.args_to_native(*ret)
-                return ivy.to_native(ret)
 
         c_args = ivy.default(constructor_args, [])
         c_kwargs = ivy.default(constructor_kwargs, {})
@@ -661,7 +426,9 @@ class ModuleConverters:
         if inspect.isclass(native_module):
             native_module = native_module(*c_args, **c_kwargs)
 
-        return TorchIvyModule(
+        from ivy.stateful.module import _TorchIvyModule
+
+        return _TorchIvyModule(
             *i_args,
             native_module=native_module,
             device=device,
