+++ ivy/stateful/module.py
@@ -1,6 +1,7 @@
 """Base class for deriving trainable modules."""
 
 # global
+from collections import OrderedDict
 import functools
 import os
 import abc
@@ -12,6 +13,7 @@ from typing import Optional, Tuple, Dict
 import ivy
 from ivy.data_classes.container import Container
 from ivy.func_wrapper import _get_first_array
+from ivy.functional.ivy.gradients import _is_variable
 from ivy.stateful.helpers import ModuleHelpers
 from ivy.stateful.converters import ModuleConverters
 
@@ -973,3 +975,233 @@ class Module(ModuleHelpers, ModuleConverters, ModuleMeta):
         if ivy.current_backend_str() == "paddle":
             loaded._convert_numpy_to_tensors()
         return loaded
+
+
+class _HaikuIvyModule(ivy.Module):
+    def __init__(self, *args, params_hk, native_module, device, devices, **kwargs):
+        self._native_module = native_module
+        self._args = args
+        self._kwargs = kwargs
+        ivy.Module.__init__(
+            self,
+            params_hk,
+            *args,
+            build_mode="on_init",
+            device=device,
+            devices=devices,
+            **kwargs,
+        )
+
+    def _create_variables(self, device, dtype):
+        return self._hk_params
+
+    def _build(self, params_hk, *args, **kwargs):
+        pass
+
+        args, kwargs = ivy.args_to_native(*args, **kwargs)
+        # noinspection PyUnresolvedReferences
+        params_dict = self._hk_flat_map_to_dict(params_hk)
+        self._hk_params = ivy.Container(params_dict, dynamic_backend=False)
+        param_iterator = self._hk_params.cont_to_iterator()
+        _, param0 = next(param_iterator, ["_", 0])
+        if hasattr(param0, "device"):
+            self._dev = ivy.as_ivy_dev(param0.device())
+        else:
+            self._dev = ivy.as_ivy_dev("cpu")
+
+    def _forward(self, *a, **kw):
+        a, kw = ivy.args_to_native(*a, **kw)
+        params_hk = self._dict_to_hk_flat_map(self.v.cont_to_dict())
+        ret = self._native_module.apply(params_hk, 0, *a, **kw)
+        if isinstance(ret, tuple):
+            return ivy.args_to_native(*ret)
+        return ivy.to_native(ret)
+
+    def _hk_flat_map_to_dict(self, hk_flat_map):
+        from haiku._src.data_structures import FlatMapping
+
+        ret_dict = dict()
+        for k, v in hk_flat_map.items():
+            new_k = k.replace("/", "|")
+            if isinstance(v, FlatMapping):
+                ret_dict[new_k] = self._hk_flat_map_to_dict(v)
+            else:
+                ret_dict[new_k] = v
+        return ret_dict
+
+    def _dict_to_hk_flat_map(self, dict_in):
+        from haiku._src.data_structures import FlatMapping
+
+        ret_flat_map = dict()
+        for k, v in dict_in.items():
+            new_k = k.replace("|", "/")
+            if isinstance(v, dict):
+                ret_flat_map[new_k] = self._dict_to_hk_flat_map(v)
+            else:
+                ret_flat_map[new_k] = v
+        return FlatMapping(ret_flat_map)
+
+
+class _FlaxIvyModule(ivy.Module):
+    def __init__(self, *args, params_fx, native_module, device, devices, **kwargs):
+        self._native_module = native_module
+        self._args = args
+        self._kwargs = kwargs
+        ivy.Module.__init__(
+            self,
+            params_fx,
+            *args,
+            build_mode="on_init",
+            device=device,
+            devices=devices,
+            **kwargs,
+        )
+
+    def _create_variables(self, device, dtype):
+        return self._fx_params
+
+    def _build(self, params_fx, *args, **kwargs):
+        import flax
+
+        args, kwargs = ivy.args_to_native(*args, **kwargs)
+        # noinspection PyUnresolvedReferences
+        params_dict = flax.core.unfreeze(params_fx)
+        self._fx_params = ivy.Container(params_dict, dynamic_backend=False)
+        param_iterator = self._fx_params.cont_to_iterator()
+        _, param0 = next(param_iterator, ["_", 0])
+        self._dev = ivy.as_ivy_dev(ivy.dev(param0))
+
+    def _forward(self, *a, **kw):
+        import flax
+
+        a, kw = ivy.args_to_native(*a, **kw)
+        params_fx = flax.core.freeze(self.v.cont_to_dict())
+        ret = self._native_module.apply(params_fx, *a, **kw)
+        if isinstance(ret, tuple):
+            return ivy.args_to_native(*ret)
+        return ivy.to_native(ret)
+
+
+class _KerasIvyModule(ivy.Module):
+    def __init__(self, *args, native_module, device, devices, **kwargs):
+        self._native_module = native_module
+        self._args = args
+        self._kwargs = kwargs
+
+        ivy.Module.__init__(self, *args, device=device, devices=devices, **kwargs)
+
+    def _create_variables(self, device=None, dtype=None):
+        return self._native_params
+
+    def _build(self, *args, **kwargs):
+        self._native_params = ivy.Container(
+            OrderedDict(
+                sorted([(param.name, param) for param in self._native_module.variables])
+            ),
+            dynamic_backend=False,
+        )
+
+    def _forward(self, *a, **kw):
+        a, kw = ivy.args_to_native(*a, **kw)
+        ret = self._native_module(*a, **kw)
+        if isinstance(ret, tuple):
+            return ivy.args_to_native(*ret)
+        return ivy.to_native(ret)
+
+
+class _PaddleIvyModule(ivy.Module):
+    def __init__(self, *args, native_module, device, devices, **kwargs):
+        self._native_module = native_module
+        self._args = args
+        self._kwargs = kwargs
+
+        ivy.Module.__init__(self, *args, device=device, devices=devices, **kwargs)
+
+    def _create_variables(self, device=None, dtype=None):
+        return self._native_params
+
+    def _build(self, *args, **kwargs):
+        self._native_params = ivy.Container(
+            OrderedDict(
+                sorted(
+                    [
+                        (k.replace(".", "/"), v)
+                        for k, v in dict(self._native_module.named_parameters()).items()
+                    ]
+                )
+            ),
+            dynamic_backend=False,
+        )
+
+    def _forward(self, *a, **kw):
+        a, kw = ivy.args_to_native(*a, **kw)
+        ret = self._native_module(*a, **kw)
+        if isinstance(ret, tuple):
+            return ivy.args_to_native(*ret)
+        return ivy.to_native(ret)
+
+
+class _TorchIvyModule(ivy.Module):
+    def __init__(self, *args, native_module, device, devices, inplace_update, **kwargs):
+        self._native_module = native_module
+        self._args = args
+        self._kwargs = kwargs
+        self._update_v = (
+            self._inplace_update_v if inplace_update else self._replace_update_v
+        )
+        ivy.Module.__init__(self, *args, device=device, devices=devices, **kwargs)
+
+    def _create_variables(self, device=None, dtype=None):
+        return self._native_params
+
+    def _build(self, *args, **kwargs):
+        self._native_params = ivy.Container(
+            OrderedDict(
+                sorted(
+                    [
+                        (k.replace(".", "/"), v)
+                        for k, v in dict(self._native_module.named_parameters()).items()
+                    ]
+                )
+            ),
+            dynamic_backend=False,
+        )
+
+    @staticmethod
+    def _inplace_update(p, v):
+        p.data = v.data
+
+    def _inplace_update_v(self, new_v):
+        ivy.Container.cont_multi_map(
+            lambda xs, kc: self._inplace_update(xs[0], xs[1]),
+            [self._native_params, new_v],
+        )
+
+    def _replace_update_v(self, new_v, native=None):
+        import torch
+
+        native = ivy.default(native, self._native_module)
+        for k, v in new_v.items():
+            if isinstance(v, ivy.Container):
+                # noinspection PyProtectedMember
+                native._modules[k] = self._replace_update_v(v, native._modules[k])
+            elif _is_variable(v):
+                # noinspection PyProtectedMember
+                native.__setattr__(k, v)
+            elif isinstance(v, torch.Tensor):
+                # noinspection PyProtectedMember
+                native.__setattr__(k, torch.nn.Parameter(v))
+            else:
+                raise ivy.utils.exceptions.IvyException(
+                    "found item in variable container {} which was neither a "
+                    "sub ivy.Container nor a variable.".format(v)
+                )
+        return native
+
+    def _forward(self, *a, **kw):
+        a, kw = ivy.args_to_native(*a, **kw)
+        self._update_v(self.v)
+        ret = self._native_module(*a, **kw)
+        if isinstance(ret, tuple):
+            return ivy.args_to_native(*ret)
+        return ivy.to_native(ret)
